change structure

+++ backend/common/src/main/java/com/common/dto/user/PointRDTO.java
@@ -4,4 +4,4 @@ import com.common.dto.structure.InterfaceDTO;
 
 import java.util.List;
 
-public record PointRDTO(Long points, String numeroPunti, List<String> nameImage) implements InterfaceDTO {}
+public record PointRDTO(Integer points, String numeroPunti, List<String> nameImage) implements InterfaceDTO {}
